requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
SQLAlchemy==2.0.9
requests==2.28.2
PyMySQL==1.0.3
Jinja2==3.1.2
python-dotenv==1.0.0
virustotal3==1.0.3
slack-sdk==3.9.0
